more check fn to seperate module
