+++ src/lib/third-party-cookies.ts
@@ -1,0 +1,36 @@
+interface TPCTestPayload {
+	tpcTest: boolean;
+	hasStorageAccess?: boolean;
+}
+
+const isTPCTestPayload = (payload: unknown): payload is TPCTestPayload =>
+	typeof payload === 'object' &&
+	payload !== null &&
+	'tpcTest' in payload &&
+	typeof payload.tpcTest === 'boolean';
+
+/**
+ * Check if third party cookies are enabled
+ * This is done by creating an iframe on another domain and checking if it can access cookies
+ **/
+const checkThirdPartyCookiesEnabled = (): void => {
+	const crossSiteIrame = document.createElement('iframe');
+	crossSiteIrame.src = `${window.guardian.config.frontendAssetsFullURL}commercial/tpc-test/index.html`;
+
+	window.addEventListener('message', ({ data }) => {
+		if (isTPCTestPayload(data)) {
+			const { hasStorageAccess } = data;
+
+			// only set targeting if the value is defined
+			if (hasStorageAccess !== undefined) {
+				window.googletag
+					.pubads()
+					.setTargeting('3pc', [hasStorageAccess ? 't' : 'f']);
+			}
+		}
+	});
+
+	document.body.appendChild(crossSiteIrame);
+};
+
+export { checkThirdPartyCookiesEnabled };
